img2text/requirements.txt
transformers==4.34.0
torch==2.0.1
Pillow==10.0.1
pymongo==4.5.0
requests==2.31.0
python-dotenv==1.0.0
tqdm==4.66.1